refactoring continues
