+++ src/components/Sidebar/Sidebar.jsx
@@ -1,5 +1,5 @@
 import * as S from './styles';
-import { Sidebar__Item } from 'components';
+import { SidebarItem } from 'components';
 import playlist01 from 'assets/img/playlist01.png';
 import playlist02 from 'assets/img/playlist02.png';
 import playlist03 from 'assets/img/playlist03.png';
@@ -7,13 +7,13 @@ import playlist03 from 'assets/img/playlist03.png';
 export const Sidebar = () => {
   return (
     <S.Sidebar>
-      <Sidebar__Item
+      <SidebarItem
         img={playlist01}
         description="day's playlist"
         loadingClass={''}
         link='/playlist/1'
       />
-      <Sidebar__Item
+      <SidebarItem
         img={playlist02}
         description="
 100 dance hits"
@@ -21,7 +21,7 @@ export const Sidebar = () => {
         link='/playlist/2'
 
       />
-      <Sidebar__Item
+      <SidebarItem
         img={playlist03}
         description="indie charge"
         loadingClass={''}
