+++ src/components/SidebarItem/SidebarItem.jsx
@@ -1,7 +1,7 @@
 import { Link } from 'react-router-dom';
 import * as S from './styles';
 
-export const Sidebar__Item = ({ img, description, loadingClass, link }) => {
+export const SidebarItem = ({ img, description, loadingClass, link }) => {
   return (
     <S.Sidebar__item className={loadingClass}>
       <Link to={link}>
