+++ src/components/index.js
@@ -10,6 +10,7 @@ export { RegBtn } from './RegBtn/RegBtn';
 export { NavLink } from './NavLink/NavLink';
 export { Nav } from './Nav/Nav';
 // Music Player
+export { MusicPlayer } from './MusicPlayer/MusicPlayer';
 export { Controls } from './MusicPlayer/components/Controls/Controls';
 export { LikeDislike } from './MusicPlayer/components/LikeDislike/LikeDislike';
 export { ProgressBar } from './MusicPlayer/components/ProgressBar/ProgressBar';
@@ -17,4 +18,13 @@ export { TrackInfo } from './MusicPlayer/components/TrackInfo/TrackInfo';
 export { Volume } from './MusicPlayer/components/Volume/Volume';
 export { Interface } from './MusicPlayer/components/Interface/Interface';
 // Sidebar
-export { Sidebar__Item } from './Sidebar__Item/Sidebar__Item';
+export { SidebarItem } from './SidebarItem/SidebarItem';
+export { Sidebar } from './Sidebar/Sidebar';
+// Tracs
+export { Tracks__FilterBtn } from './Tracks__FilterBtn/Tracks__FilterBtn';
+export { Tracks__SearchFilter } from './Tracks__SearchFilter/Tracks__SearchFilter';
+export { Tracks__FilterModal } from './Tracks__FilterModal/Tracks__FilterModal';
+export { Tracks__Track } from './Tracks__Track/Tracks__Track';
+export { Tracks__ListHead } from './Tracks__ListHead/Tracks__ListHead';
+export { Tracks__List } from './Tracks__List/Tracks__List';
+
