+++ src/pages/MainPage/styles.js
